MRZlib: merge zlib-ng compress path back into MRZlib.cpp
Drop the separate MRZlibNg.cpp; fold the zng_-prefixed
zlibCompressStream overloads directly into MRZlib.cpp alongside
zlibDecompressStream.

Both <zlib.h> and <zlib-ng.h> include cleanly in the same TU: the Z_*
return-code / flush-mode / strategy constants are defined in both
headers with identical values (C preprocessor allows a redefinition
with the same replacement list, no warning), and the function names
(deflate/inflate vs zng_deflate/zng_inflate) and struct types
(z_stream vs zng_stream) are disjoint. The shared zlibToString and
windowBitsFor helpers are reused across compress (zng_*) and
decompress (stock zlib) since MAX_WBITS equals Z_MAX_WINDOWBITS = 15
and the Z_* code values match.

No functional change relative to the previous commit on this branch;
just collapses two TUs into one.
