+++ source/MRMesh/MRZlib.cpp
@@ -2,9 +2,18 @@
 #include "MRBuffer.h"
 #include "MRFinally.h"
 
+// zlibCompressStream uses zlib-ng in native mode (zng_ prefix, libz-ng).
+// zlibDecompressStream stays on stock zlib. Both headers are included here
+// because they share the Z_* return-code / flush-mode / strategy constants
+// with identical values -- a redefinition with the same replacement list is
+// allowed by the C preprocessor and does not produce a warning. The function
+// names (deflate/inflate vs zng_deflate/zng_inflate) and struct types
+// (z_stream vs zng_stream) are disjoint, so both libraries coexist cleanly.
 #include <zlib.h>
+#include <zlib-ng.h>
 
 #include <cassert>
+#include <cstdint>
 
 namespace
 {
@@ -59,11 +68,66 @@ int windowBitsFor( bool rawDeflate )
 namespace MR
 {
 
-// zlibCompressStream(...) and its int-level overload are implemented in
-// MRZlibNg.cpp against zlib-ng's native API (zng_* prefix). Decompression
-// stays on stock zlib -- zlibDecompressStream has to accept bytes written
-// by any deflate producer, so compatibility matters more than speed, and
-// libzip still uses stock zlib as well.
+Expected<void> zlibCompressStream( std::istream& in, std::ostream& out, const ZlibCompressParams& params )
+{
+    Buffer<char> inChunk( cChunkSize ), outChunk( cChunkSize );
+    zng_stream stream{};
+    int ret;
+    if ( Z_OK != ( ret = zng_deflateInit2( &stream, params.level, Z_DEFLATED,
+                                           windowBitsFor( params.rawDeflate ),
+                                           kDefaultMemLevel, Z_DEFAULT_STRATEGY ) ) )
+        return unexpected( zlibToString( ret ) );
+
+    MR_FINALLY {
+        zng_deflateEnd( &stream );
+    };
+
+    if ( params.stats )
+        *params.stats = {};
+
+    while ( !in.eof() )
+    {
+        in.read( inChunk.data(), static_cast<std::streamsize>( inChunk.size() ) );
+        if ( in.bad() )
+            return unexpected( "I/O error" );
+        stream.next_in = reinterpret_cast<uint8_t*>( inChunk.data() );
+        stream.avail_in = static_cast<unsigned>( in.gcount() );
+        assert( stream.avail_in <= static_cast<unsigned>( inChunk.size() ) );
+
+        if ( params.stats )
+        {
+            params.stats->crc32 = static_cast<uint32_t>(
+                zng_crc32( params.stats->crc32, stream.next_in, stream.avail_in ) );
+            params.stats->uncompressedSize += stream.avail_in;
+        }
+
+        const int flush = in.eof() ? Z_FINISH : Z_NO_FLUSH;
+        do
+        {
+            stream.next_out = reinterpret_cast<uint8_t*>( outChunk.data() );
+            stream.avail_out = static_cast<unsigned>( outChunk.size() );
+            ret = zng_deflate( &stream, flush );
+            if ( Z_OK != ret && Z_STREAM_END != ret )
+                return unexpected( zlibToString( ret ) );
+
+            assert( stream.avail_out <= static_cast<unsigned>( outChunk.size() ) );
+            const unsigned written = static_cast<unsigned>( outChunk.size() ) - stream.avail_out;
+            out.write( outChunk.data(), written );
+            if ( out.bad() )
+                return unexpected( "I/O error" );
+            if ( params.stats )
+                params.stats->compressedSize += written;
+        }
+        while ( stream.avail_out == 0 );
+    }
+
+    return {};
+}
+
+Expected<void> zlibCompressStream( std::istream& in, std::ostream& out, int level )
+{
+    return zlibCompressStream( in, out, ZlibCompressParams{ .level = level } );
+}
 
 Expected<void> zlibDecompressStream( std::istream& in, std::ostream& out, const ZlibParams& params )
 {
